+++ platform/variables.tf
@@ -160,6 +160,18 @@ variable "postgresql_password" {
   sensitive   = true
 }
 
+variable "keycloak_db_password" {
+  description = "Keycloak database password"
+  type        = string
+  sensitive   = true
+}
+
+variable "airflow_db_password" {
+  description = "Airflow database password"
+  type        = string
+  sensitive   = true
+}
+
 variable "postgresql_username" {
   description = "PostgreSQL application database owner username"
   type        = string